Tier 1 and 2 of the patch list: seven more, five of them reproduced first
Of the nineteen strategies the adversary panel returned, eight had been
tested. This does the remaining eleven that were worth testing. Five
reproduced against the real contracts, two more were confirmed by reading
the source, and all seven are fixed.

  · revokeSession was `delete sessionOf[key]`, which cannot reach a mapping
    — so sessionTarget and sessionSelector survived it. Revocation looked
    total and re-granting the key brought back EVERY permission it had ever
    held. Grant [poolA], revoke, re-grant [poolB], still reaches poolA. Both
    paths now bump a per-key epoch, which retires the old entries in
    constant gas.

  · The approval defence was a six-selector enumeration — the exact shape
    this codebase's own argument says cannot be completed — and Permit2's
    approve(address,address,uint160,uint48) walked through it with identical
    standing custody. Inverted: while sealed, a call to a MANIFEST asset must
    carry transfer or transferFrom and nothing else. Every approval shape,
    every permit shape, and words nobody has invented yet are refused by one
    rule without naming any of them. Calls to anything not promised stay
    unrestricted, because an approval on an asset nobody was promised is not
    the seal's business.

    The cost is real and is what "default" means: a holder who wants to
    claim() on a promised asset while sealed cannot. Two tests that
    demonstrated "a sealed vault can still act" were doing it by calling a
    manifest asset, which is now precisely the restricted case; they act on
    something unpromised instead, and a new assertion states the cost.

  · balanceOf is the same WORD for ERC-20 and ERC-721 and not the same FACT.
    A sealed vault swapped a valuable NFT for a worthless one through a
    venue — one out, one in, count unmoved, nothing to measure. guardNFT
    names a piece and ownerOf answers exactly. The test carries a control
    that swaps freely with only the collection guarded, so the fix is
    credited to the mechanism that actually catches it rather than to
    whichever guard happened to fire.

  · Refusing a call that ends with an asset unreadable is correct, and was a
    door a manifest asset could shut at will: a token whose balanceOf
    reverts on its own condition made every sealed call revert, and a
    reverted call never persists, so the trap re-armed for the seal's full
    length. An asset the account cannot read may now be unguarded even while
    sealed — the seal could never promise about it, and unmeasurable() had
    been saying so publicly the whole time.

  · Attestations had no nonce and no deadline, so one signature
    authenticated the same statement to everyone forever. Both are in the
    signed struct now, plus retireAttestations() to void every signature the
    account ever gave.

  · deposit re-checked both sides against maxDeposit, so once trading pushed
    a reserve past the cap — swap never consults it — a one-wei quote-only
    top-up was refused on account of a base reserve it had not touched.
    Under a live bond that left the holder with no operable function at all.

  · The anchored offsets were a uint112 holding up to eight times a uint112.
    I could not drive a reserve high enough to truncate through the shipped
    caps, so this is a latent hazard rather than a reproduced bug. A cast
    that can silently produce a different curve than the one committed is
    worth four bytes to remove rather than an argument to have. uint128.

Two claims refuted on the first attempt turned out to be my fixtures being
wrong, not the claims: one mock sent the NFT to msg.sender, which was the
vault itself, and one drove the reserve with a single trade the output cap
refuses. Both are recorded. A refutation from a broken fixture is worse than
no test, because it reads like an answer.

  421 assertions · 14 properties · 15 claims refuted · 220 ticks green

+++ test/Pool.t.sol
@@ -86,7 +86,7 @@ contract PoolTest is Test {
     ///      conserves — and a round trip could extract the difference. The
     ///      offsets are anchored now; this reads what was anchored.
     function _k() internal view returns (uint256) {
-        (, , uint112 rb, uint112 rq, , , , , uint112 vb, uint112 vq) = pool.marketOf(1);
+        (, , uint112 rb, uint112 rq, , , , , uint128 vb, uint128 vq) = pool.marketOf(1);
         return (uint256(rb) + vb) * (uint256(rq) + vq);
     }
